antigravidade/backend/requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
faster-whisper
docxtpl
python-multipart
